# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BAM Store API",
        "description": "API Documentation for BAM Store",
        "contact": {
            "email": "admin@bamstore.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/admin/categories": {
            "get": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Get list of categories (Admin)",
                "operationId": "c9d370a5ffd5b98edbb40db5be419cee",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Create a new category",
                "operationId": "41138cf423f5e7f784ea5dd54c193994",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Electronics"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/categories/{id}": {
            "get": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Get category details",
                "operationId": "f033e9ad2dbf8325ecdbe9823c15749a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID or Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Update a category",
                "operationId": "21b9db60411febf91a3df563499799fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID or Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Electronics"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "updated-electronics"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Delete a category",
                "operationId": "6e07250e2afdb687f81f9d036eb5f43f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID or Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Category deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/admin/orders": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Get list of orders (Admin)",
                "operationId": "f0a30422ae39690543a852d0833c2115",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "new",
                                "paid",
                                "processed",
                                "ready_pickup",
                                "completed",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "qobilah",
                        "in": "query",
                        "description": "Filter by qobilah",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/orders/{id}": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Get order details (Admin)",
                "operationId": "9bb065c37d5443718dde4f93e9e1bf8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID or Order Number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/orders/{id}/status": {
            "put": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Update order status",
                "operationId": "b1a003d5032f6ea5842b79836c5c97e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID or Order Number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "new",
                                            "paid",
                                            "processed",
                                            "ready_pickup",
                                            "completed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order status updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}